Moved physical interfaces. Small updates to structure materials
