+++ Physical_oM/Elements/Column.cs
@@ -26,7 +26,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-using BH.oM.Physical.Interface;
+using BH.oM.Physical.Properties;
 using BH.oM.Base;
 using BH.oM.Geometry;
