+++ Physical_oM/Properties/Construction/IConstruction.cs
@@ -1,6 +1,6 @@
 ﻿/*
  * This file is part of the Buildings and Habitats object Model (BHoM)
- * Copyright (c) 2015 - 2019, the respective contributors. All rights reserved.
+ * Copyright (c) 2015 - 2018, the respective contributors. All rights reserved.
  *
  * Each contributor holds copyright over their respective contributions.
  * The project versioning (Git) records all such contribution source information.
@@ -20,24 +20,23 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BH.oM.Structure.Properties.Materials
+using BH.oM.Base;
+
+namespace BH.oM.Physical.Properties.Construction
 {
-    public interface IIsotropicMaterial : IStructuralMaterial
+    public interface IConstruction : IBHoMObject
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
-        double YoungsModulus { get; set; }
-        double PoissonsRatio { get; set; }
-        double ThermalExpansionCoeff { get; set; }
+
 
         /***************************************************/
     }
