+++ Physical_oM/Properties/IFramingElementProperty.cs
@@ -27,22 +27,16 @@ using System.Text;
 using System.Threading.Tasks;
 
 using BH.oM.Base;
-using BH.oM.Geometry;
-using BH.oM.Physical.Properties.Construction;
-using BH.oM.Physical.Elements;
 
-namespace BH.oM.Physical.Interface
+namespace BH.oM.Physical.Properties
 {
-    public interface ISurface : IBHoMObject, IPhysical
+    public interface IFramingElementProperty : IBHoMObject
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
-        BH.oM.Geometry.ISurface OuterBoundary { get; set; }
-        IConstruction Construction { get; set; }
-        List<IOpening> Openings { get; set; }
-        Offset Offset { get; set; }
+
 
         /***************************************************/
     }
