+++ Physical_oM/Properties/IMaterialProperties.cs
@@ -1,6 +1,6 @@
 ﻿/*
  * This file is part of the Buildings and Habitats object Model (BHoM)
- * Copyright (c) 2015 - 2019, the respective contributors. All rights reserved.
+ * Copyright (c) 2015 - 2018, the respective contributors. All rights reserved.
  *
  * Each contributor holds copyright over their respective contributions.
  * The project versioning (Git) records all such contribution source information.
@@ -20,26 +20,23 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using BH.oM.Geometry;
 
-namespace BH.oM.Structure.Properties.Materials
+using BH.oM.Base;
+
+namespace BH.oM.Physical.Properties
 {
-    public interface IOrthotropicMaterial : IStructuralMaterial
+    public interface IMaterialProperties : IBHoMObject
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
-        Vector YoungsModulus { get; set; }
-        Vector PoissonsRatio { get; set; }
-        Vector ThermalExpansionCoeff { get; set; }
-        Vector ShearModulus { get; set; }
+
 
         /***************************************************/
     }
